updating with useDispatch
