+++ ui/pages/settings/networks-tab/network-list/network-list.js
@@ -4,14 +4,13 @@ import classnames from 'classnames';
 import NetworkListItem from '../network-list-item';
 
 const NetworkList = ({
-  networksToRender,
-  selectedNetwork,
-  setSelectedSettingsRpcUrl,
-  networkIsSelected,
-  networkDefaultedToProvider,
-  providerUrl,
-  providerType,
   isFullScreen,
+  networkIsSelected,
+  networksToRender,
+  networkDefaultedToProvider,
+  providerType,
+  providerUrl,
+  selectedRpcUrl,
 }) => {
   return (
     <div
@@ -24,9 +23,7 @@ const NetworkList = ({
         <NetworkListItem
           key={`settings-network-list:${network.rpcUrl}`}
           network={network}
-          selectedNetwork={selectedNetwork}
-          selectRpcUrl={selectedNetwork.rpcUrl}
-          setSelectedSettingsRpcUrl={setSelectedSettingsRpcUrl}
+          selectedRpcUrl={selectedRpcUrl}
           networkIsSelected={networkIsSelected}
           providerUrl={providerUrl}
           providerType={providerType}
@@ -38,14 +35,13 @@ const NetworkList = ({
 };
 
 NetworkList.propTypes = {
-  networksToRender: PropTypes.arrayOf(PropTypes.object).isRequired,
-  selectedNetwork: PropTypes.object,
-  setSelectedSettingsRpcUrl: PropTypes.func.isRequired,
-  networkIsSelected: PropTypes.bool,
-  networkDefaultedToProvider: PropTypes.bool,
-  providerUrl: PropTypes.string,
-  providerType: PropTypes.string,
   isFullScreen: PropTypes.bool.isRequired,
+  networkDefaultedToProvider: PropTypes.bool,
+  networkIsSelected: PropTypes.bool,
+  networksToRender: PropTypes.arrayOf(PropTypes.object).isRequired,
+  providerType: PropTypes.string,
+  providerUrl: PropTypes.string,
+  selectedRpcUrl: PropTypes.string,
 };
 
 export default NetworkList;
