+++ ui/pages/settings/networks-tab/network-tab.js
@@ -1,12 +1,10 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { useHistory } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
 import { useI18nContext } from '../../../hooks/useI18nContext';
-import {
-  ADD_NETWORK_ROUTE,
-  NETWORKS_ROUTE,
-  DEFAULT_ROUTE,
-} from '../../../helpers/constants/routes';
+import { ADD_NETWORK_ROUTE } from '../../../helpers/constants/routes';
+import { setSelectedSettingsRpcUrl } from '../../../store/actions';
 import Button from '../../../components/ui/button';
 import NetworkForm from './network-form';
 import NetworkTabContent from './network-tab-content';
@@ -43,64 +41,45 @@ SubHeader.propTypes = {
 };
 
 const NetworkTab = ({
-  editRpc,
+  addNewNetwork,
+  isFullScreen,
+  networkDefaultedToProvider,
   networkIsSelected,
   networksToRender,
-  selectedNetwork,
-  setRpcTarget,
-  setSelectedSettingsRpcUrl,
-  showConfirmDeleteNetworkModal,
-  providerUrl,
   providerType,
-  networkDefaultedToProvider,
+  providerUrl,
+  selectedNetwork,
   shouldRenderNetworkForm,
-  isFullScreen,
-  setNewNetworkAdded,
-  addNewNetwork,
 }) => {
   const t = useI18nContext();
-  const history = useHistory();
+  const dispatch = useDispatch();
+
   useEffect(() => {
     return () => {
-      setSelectedSettingsRpcUrl('');
+      dispatch(setSelectedSettingsRpcUrl(''));
     };
-  }, [setSelectedSettingsRpcUrl]);
+  }, [dispatch]);
+
   return (
     <div className="networks-tab__body">
       {isFullScreen ? <SubHeader addNewNetwork={addNewNetwork} /> : null}
       <div className="networks-tab__content">
         {addNewNetwork ? (
           <NetworkForm
-            setRpcTarget={setRpcTarget}
-            onClear={(shouldUpdateHistory = true) => {
-              if (shouldUpdateHistory) {
-                history.push(NETWORKS_ROUTE);
-              }
-            }}
-            onAddNetwork={() => {
-              history.push(DEFAULT_ROUTE);
-            }}
-            rpcPrefs={selectedNetwork.rpcPrefs}
             networksToRender={networksToRender}
-            setNewNetworkAdded={setNewNetworkAdded}
             addNewNetwork={addNewNetwork}
-            history={history}
           />
         ) : (
           <>
             <NetworkTabContent
-              setRpcTarget={setRpcTarget}
-              showConfirmDeleteNetworkModal={showConfirmDeleteNetworkModal}
-              setSelectedSettingsRpcUrl={setSelectedSettingsRpcUrl}
-              selectedNetwork={selectedNetwork}
-              editRpc={editRpc}
-              providerUrl={providerUrl}
-              providerType={providerType}
-              networkDefaultedToProvider={networkDefaultedToProvider}
-              networksToRender={networksToRender}
               isFullScreen={isFullScreen}
-              shouldRenderNetworkForm={shouldRenderNetworkForm}
+              networkDefaultedToProvider={networkDefaultedToProvider}
               networkIsSelected={networkIsSelected}
+              networksToRender={networksToRender}
+              providerType={providerType}
+              providerUrl={providerUrl}
+              selectedNetwork={selectedNetwork}
+              shouldRenderNetworkForm={shouldRenderNetworkForm}
             />
             {!isFullScreen && !shouldRenderNetworkForm ? (
               <div className="networks-tab__networks-list-popup-footer">
@@ -123,20 +102,15 @@ const NetworkTab = ({
 };
 
 NetworkTab.propTypes = {
-  editRpc: PropTypes.func.isRequired,
+  addNewNetwork: PropTypes.bool,
+  isFullScreen: PropTypes.bool.isRequired,
+  networkDefaultedToProvider: PropTypes.bool,
   networkIsSelected: PropTypes.bool,
   networksToRender: PropTypes.arrayOf(PropTypes.object).isRequired,
-  selectedNetwork: PropTypes.object,
-  setRpcTarget: PropTypes.func.isRequired,
-  setSelectedSettingsRpcUrl: PropTypes.func.isRequired,
-  showConfirmDeleteNetworkModal: PropTypes.func.isRequired,
-  providerUrl: PropTypes.string,
   providerType: PropTypes.string,
-  networkDefaultedToProvider: PropTypes.bool,
+  providerUrl: PropTypes.string,
+  selectedNetwork: PropTypes.object,
   shouldRenderNetworkForm: PropTypes.bool.isRequired,
-  isFullScreen: PropTypes.bool.isRequired,
-  setNewNetworkAdded: PropTypes.func.isRequired,
-  addNewNetwork: PropTypes.bool,
 };
 
 export default NetworkTab;
